refactor to add unit test to ensure there is coverage in the schema
